% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feeds.R
\name{get_feeds}
\alias{get_feeds}
\title{Get Feeds}
\usage{
get_feeds(
  limit = 10,
  offset = 0,
  status = NULL,
  provider = NULL,
  producer_url = NULL,
  is_official = NULL,
  base_url = "https://api.mobilitydatabase.org/v1",
  access_token = getOption("MOBILITY_FEED_API_ACCESS_TOKEN")
)
}
\arguments{
\item{limit}{\code{Integer.} The number of items to be returned.}

\item{offset}{\code{Integer.} Offset of the first item to return.}

\item{provider}{\code{Character.} List only feeds with the specified value. Can be a partial match. Case insensitive.}

\item{producer_url}{\code{Character.} List only feeds with the specified value. Can be a partial match. Case insensitive.}

\item{is_official}{\code{Logical.} If true, only return official feeds.}

\item{base_url}{\code{Character.} URL of mobility database API.}

\item{access_token}{\code{Character.} Access/bearer token. Can be obtained with \code{\link[=get_token]{get_token()}} and will be stored as option \code{MOBILITY_FEED_API_ACCESS_TOKEN}.}
}
\value{
A \code{tibble} of mobility data results.
}
\description{
Get feeds from the Mobility Database.
}
