% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs_feeds_by_id.R
\name{get_gtfs_feed_related_datasets}
\alias{get_gtfs_feed_related_datasets}
\title{Get related datasets}
\usage{
get_gtfs_feed_related_datasets(
  id,
  base_url = "https://api.mobilitydatabase.org/v1",
  access_token = getOption("MOBILITY_FEED_API_ACCESS_TOKEN")
)
}
\arguments{
\item{id}{\code{Character.} The ID of the requested dataset.}

\item{base_url}{\code{Character.} URL of mobility database API.}

\item{access_token}{\code{Character.} Access/bearer token. Can be obtained with \code{\link[=get_token]{get_token()}} and will be stored as option \code{MOBILITY_FEED_API_ACCESS_TOKEN}.}
}
\value{
A \code{tibble} of mobility data results.
}
\description{
Get datasets related to specified GTFS feed
}
