% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs_rt_feeds.R
\name{get_gtfs_rt_feeds}
\alias{get_gtfs_rt_feeds}
\title{Get real-time GTFS Feeds}
\usage{
get_gtfs_rt_feeds(
  limit = 10,
  offset = 0,
  status = NULL,
  provider = NULL,
  producer_url = NULL,
  country_code = NULL,
  entity_types = NULL,
  subdivision_name = NULL,
  municipality = NULL,
  is_official = NULL,
  base_url = "https://api.mobilitydatabase.org/",
  access_token = getOption("MOBILITY_FEED_API_ACCESS_TOKEN")
)
}
\arguments{
\item{limit}{\code{Integer.} The number of items to be returned.}

\item{offset}{\code{Integer.} Offset of the first item to return.}

\item{provider}{\code{Character.} List only feeds with the specified value. Can be a partial match. Case insensitive.}

\item{producer_url}{\code{Character.} List only feeds with the specified value. Can be a partial match. Case insensitive.}

\item{country_code}{\code{Character.} Filter feeds by their exact country code.}

\item{subdivision_name}{\code{Character.} List only feeds with the specified value (e.g. "Oregon", "Niedersachsen"). Can be a partial match. Case insensitive.}

\item{municipality}{\code{Character.} List only feeds with the specified value. Can be a partial match. Case insensitive.}

\item{is_official}{\code{Logical.} If true, only return official feeds.}

\item{base_url}{\code{Character.} URL of mobility database API.}

\item{access_token}{\code{Character.} Access/bearer token. Can be obtained with \code{\link[=get_token]{get_token()}} and will be stored as option \code{MOBILITY_FEED_API_ACCESS_TOKEN}.}
}
\value{
A \code{tibble} of mobility data results.
}
\description{
Get real-time GTFS feeds from the Mobility Database.
}
