% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get Metadata}
\usage{
get_metadata(
  base_url = "https://api.mobilitydatabase.org/v1",
  access_token = getOption("MOBILITY_FEED_API_ACCESS_TOKEN")
)
}
\arguments{
\item{base_url}{\code{Character.} URL of mobility database API.}

\item{access_token}{\code{Character.} Access/bearer token. Can be obtained with \code{\link[=get_token]{get_token()}} and will be stored as option \code{MOBILITY_FEED_API_ACCESS_TOKEN}.}
}
\value{
A \code{tibble} of mobility data results.
}
\description{
Get metadata from the Mobility Database.
}
