% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search.R
\name{get_search}
\alias{get_search}
\title{Search feeds on feed name, location and provider's information.}
\usage{
get_search(
  search_query = NULL,
  feature = NULL,
  limit = 10,
  offset = 0,
  status = NULL,
  feed_id = NULL,
  data_type = NULL,
  is_official = NULL,
  version = NULL,
  base_url = "https://api.mobilitydatabase.org/",
  access_token = getOption("MOBILITY_FEED_API_ACCESS_TOKEN")
)
}
\arguments{
\item{search_query}{\code{Character.} General search query to match against transit provider, location, and feed name.}

\item{feature}{\code{Character.} Filter feeds by their GTFS features. \href{https://gtfs.org/getting-started/features/overview/}{GTFS features definitions defined here}.}

\item{limit}{\code{Integer.} The number of items to be returned.}

\item{offset}{\code{Integer.} Offset of the first item to return.}

\item{status}{description Filter feeds by their status. Available values: active, \code{deprecated}, \code{inactive}, \code{development}, \code{future}.}

\item{data_type}{Comma separated list of data types to filter by. Valid values are \code{gtfs}, \code{gtfs_rt} and \code{gbfs}.}

\item{is_official}{\code{Logical.} If true, only return official feeds.}

\item{version}{\code{Character.} Comma separated list of GBFS versions to filter by.}

\item{base_url}{\code{Character.} URL of mobility database API.}

\item{access_token}{\code{Character.} Access/bearer token. Can be obtained with \code{\link[=get_token]{get_token()}} and will be stored as option \code{MOBILITY_FEED_API_ACCESS_TOKEN}.}
}
\value{
A \code{tibble} of mobility data results.
}
\description{
The current implementation leverages the text search functionalities from PostgreSQL, in particularly \code{plainto_tsquery}.
}
