% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Obtain bearer token}
\usage{
get_token(refresh_token = Sys.getenv("MOBILITY_FEED_API_REFRESH_TOKEN"))
}
\arguments{
\item{refresh_token}{Character. Refresh token can be obtained in user area
at \url{https://mobilitydatabase.org/}.}
}
\description{
Provided a refresh token will return a bearer token and store it as option
\code{MOBILITY_FEED_API_ACCESS_TOKEN}.
}
