% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilityfeed-package.R
\docType{package}
\name{mobilityfeed-package}
\alias{mobilityfeed}
\alias{mobilityfeed-package}
\title{mobilityfeed}
\description{
Wrapper for the Mobility Database Catalog API. The \href{https://mobilitydatabase.org/}{Mobility Database} aggregates data feeds for public transit mobility data including GTFS feeds, real-time GTFS feeds, and GBFS feeds.
}
\keyword{internal}
